Running logo/demo.py

+++ logo/demo.py
@@ -4,9 +4,9 @@ from tealight.logo import (move,
 
 colors = ["red", "green", "blue"]
 
-r=100
+r=100.0
 
 for i in range(0,r):
   move(2)
-  turn(360/r)
+  turn(float(360.0/r))
   color(colors[i%3])
